pyproject.toml
[tool.poetry]
name = "evalprompt"
version = "0.1.0"
description = ""
authors = ["wenyutang-ms <wenyutang@microsoft.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.58.1"
azure-ai-evaluation = "^1.1.0"
pandas = "^2.2.3"
ipykernel = "^6.29.5"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
